+++ src/market_info.py
@@ -90,10 +90,15 @@ def get_purchased_sold_cards(account_name, start_date, end_date):
             df1 = pd.DataFrame({'spl_id': json.loads(operation['json'])['items']})
             sm_market_purchase = pd.concat([sm_market_purchase, df1])
         elif operation['id'] == 'sm_sell_cards':
-            cards = json.loads(operation['json'])
-            for card in cards:
-                df1 = pd.DataFrame({'card': card['cards']})
+            card_op = json.loads(operation['json'])
+            if isinstance(card_op, dict):
+                df1 = pd.DataFrame({'card': card_op['cards']})
                 potential_sell = pd.concat([potential_sell, df1])
+            else:
+                for card in card_op:
+                    df1 = pd.DataFrame({'card': card['cards']})
+                    potential_sell = pd.concat([potential_sell, df1])
+
 
     # process purchases
     purchases = pd.DataFrame()